Delete comments by product and user id

+++ ECommerce.WebApp/Utils/ManageFiles.cs
@@ -44,30 +44,36 @@ namespace ECommerce.WebApp.Utils
         }
         public void DeleteFiles(List<string> fileNames, string path)
         {
-            string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath, path);
-            // remove previous image
-            DirectoryInfo uploadDirectory = new DirectoryInfo(uploadsFolder);
-            foreach (var fileName in fileNames)
+            if (fileNames != null && fileNames.Count > 0)
             {
-                foreach (FileInfo file in uploadDirectory.GetFiles())
+                string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath, path);
+                // remove previous image
+                DirectoryInfo uploadDirectory = new DirectoryInfo(uploadsFolder);
+                foreach (var fileName in fileNames)
                 {
-                    if (file.Name == fileName)
+                    foreach (FileInfo file in uploadDirectory.GetFiles())
                     {
-                        file.Delete();
+                        if (file.Name == fileName)
+                        {
+                            file.Delete();
+                        }
                     }
                 }
             }
         }
         public void DeleteFile(string fileName, string path)
         {
-            string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath, path);
-            // remove previous image
-            DirectoryInfo uploadDirectory = new DirectoryInfo(uploadsFolder);
-            foreach (FileInfo file in uploadDirectory.GetFiles())
+            if (!String.IsNullOrEmpty(fileName))
             {
-                if (file.Name == fileName)
+                string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath, path);
+                // remove previous image
+                DirectoryInfo uploadDirectory = new DirectoryInfo(uploadsFolder);
+                foreach (FileInfo file in uploadDirectory.GetFiles())
                 {
-                    file.Delete();
+                    if (file.Name == fileName)
+                    {
+                        file.Delete();
+                    }
                 }
             }
         }
